I made tallyRowHeuristic use foldl', wrote a tool to run tallyRowHeuristic on the first n lines of a file (with the idea that you can run that operation once for a given automatically updated CSV datablock in the pipeline with the assumption that the column types won't change too much), and also just made Read be derived for ProtoCellType.
Anywayyyys, so like, the choice between foldr and foldl' is a matter of when
you need your base value.  foldr uses it when it gets to the end of the input
data structure, and foldl' uses it at the beginning.

In order to do addition, you need two numbers, so foldr must fully evaluate
the input structure at the start of its run.

In order to do cons, you need a head element and that's about it, so foldr
stays lazy on the input list.  So like, a lazy linked list produced by foldr
is going to be the head of the list followed by another call to foldr.

So yeah, pretty much, any time you're not putting out a lazy data structure,
just use foldl', which is just good old fashioned tail recursion; i.e. a loop.